Do not start JmxAgent by default
`JmxAgent` responsibility is to expose (or actually: make JVM expose)
the platform MBeanServer over JMX protocol. The purpose of enabling it
by default on some port (even when no port was configured) was to make
sure a JMX connection is always available. Such connection can be still
utilized without knowing the port number by using a "JMX discovery"
agent on the host.

Starting JMX agent unconditionally has however two problems. Exposing
the platform MBeanServer to the network may be a security issue in
certain scenarios. For this reason, the functionality is removed.

If, in the future, we want to support JMX agent exposure on
automatically chosen port, it needs to be enabled explicitly in the
configuration (e.g.  "auto" configuration option or equivalent). Then,
we would also need to fix the technical aspect of choosing a port
number, since the logic (`NetUtils.findUnusedPort`) being removed in
this commit was found unreliable in certain environments (e.g. Presto
product tests -- an Airlift application running inside a docker
container on contemporary GitHub Actions environment).